New version to play nice with their cloud protection
If their site suspects you of being a bot, it issues a challenge from a service called Incapsula.

I'm dealing with this by saving all the cookies that the site sets, since I think that might help the site remember who we are if the user passes this check during login.

I also detect this error and try calling lost_authentication. I don't know if it will work.

Annoyingly, there is a bug with the pagination on the IndieGala site which is preventing my library from loading. My theory is that it breaks when it needs to paginate more than 10 items.